Restructure of directory
