+++ tests/test.py
@@ -3,11 +3,14 @@ import argparse
 import sys
 import os
 import inspect
-import mei2volpiano
 
 current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
 parent_dir = os.path.dirname(current_dir)
 sys.path.insert(0, parent_dir)
+
+import src.mei2volpiano as mei2volpiano
+
+
 
 # import doctest
